requirements.txt
Django~=4.2.9
pymupdf~=1.23.14
python-dotenv==1.0.1

openai~=1.2.4
pathlib~=1.0.1
tiktoken==0.5.2

numpy==1.26.2
pandas==1.5.3
pymongo==4.6.1
weaviate-client==4.5.0
python-social-auth==0.3.6
social-auth-app-django==5.4.0
Office365-REST-Python-Client==2.3.16

boto3==1.34.111
django-storages==1.14.3
django-cors-headers==4.3.1
django-widget-tweaks==1.5.0

gunicorn==21.2.0
whitenoise==6.6.0

mysql-connector-python==9.0.0